requirements.txt
crewai
crewai_tools
python-dotenv
streamlit

